move component files to sections
