+++ src/sections/Projects/index.js
@@ -7,10 +7,10 @@ import Container from 'semantic-ui-react/dist/commonjs/elements/Container';
 import Header from 'semantic-ui-react/dist/commonjs/elements/Header';
 import Divider from 'semantic-ui-react/dist/commonjs/elements/Divider';
 
-import { sortedProjects } from '../selectors/projects';
+import { sortedProjects } from '../../selectors/projects';
 
-import NewProject from '../components/NewProject';
-import ProjectsList from '../components/ProjectsList';
+import NewProject from './components/NewProject/index';
+import ProjectsList from './components/ProjectsList/index';
 
 function Projects({ projects }: { projects: Array<projectTreeType> }) {
   return (
